Route mrview capture progress through cancellable ProgressBar
The mrview screen capture tool's ad hoc multi-frame progress dialog
is replaced with MR::ProgressBar, and its files are reorganised into
a dedicated screen_capture subdirectory and namespace. Adding a cancel
button required extending ProgressBar: a first attempt latched a global
set_next_cancellable() flag ahead of construction, but analysis showed
the flag could leak onto an unrelated dialog if a ProgressBar was
suppressed or never displayed, so it was reverted.  The final design
instead makes ProgressBar's destructor virtual and lets GUI code define
a Cancellable subclass that Dialog::ProgressBar::display() detects via
dynamic_cast, keeping the cancelled state local to each instance with
no concept of cancellation leaking into the base class's interface.

Prompts:
1. > 1. Claude has made use of a QProgressDialog in the mrview Screenshot tool for tracking progress of a multi-frame capture. Check the MRtrix3 C++ GUI code base for precedent on this. User believes that MRtrix3 class ProgressBar is intended for use even in GUI applications; modify if correct.
   > 2. New features required for super-sampling have been placed in "cpp/gui/mrview/capture_buffer.*". Assess what fraction of this functionality is used exclusively by the screen capture tool. If volume of code in these files that is specific to the screen capture tool exceeds 50% of total content, do the following steps; otherwise skip to next enumerated point. Make new directory cpp/gui/mrview/tool/screen_capture/. Move cpp/gui/mrview/tool/screen_capture.* into cpp/gui/mrview/tool/screen_capture/. Move cpp/gui/mrview/capture_buffer.* into cpp/gui/mrview/tool/screen_capture/. Place that code in MRView::Tool namespace. Restore compilation. Any code within these files accessed by any MRView code outside of the screen capture tool will require refactoring; present summary of any such code to user and prompt for further instruction.
   > 3. Resolve recently added capability of the screen capture toolbar to generate images with MSAA against pre-existing MRtrix config file key "MSAA". Evaluate whether that config file option is functional, and whether its mechanism of operation can be reconciled against the recent changes that were made only considering the context of the screen capture tool.
   Questions:
   > Q: Window::captureGL() (core window.cpp) now depends on Tool::CaptureBuffer, nested under the screen_capture tool directory/namespace. How should this coupling be resolved?
   > A: Move captureGL() into the tool
2. > Despite identification of the MR::ProgressBar precedent in GUI applications, Claude has still nevertheless used custom code rather than simply instantiating an MR::ProgressBar. Claude gives some justification for this. It is however not clear whether those context-specific changes were made because of context-specific reasons, or whether these are more general issues with the use of MR::ProgressBar in GUI applications for which Claude has implemented a local rather than global fix. Describe exactly what justifies use of custom progress window handling in the screen capture tool.
3. > A cancel functionality for MR::ProgressBar does not make sense in any terminal application. It may however make sense in more than one context in GUI applications. It would also be preferable to preserve the delay before appearance behaviour in this context. Additionally, with the addition of a progress bar to the capture process, it would make sense that the stop control would be moved from the toolbar to the progress dialog window, so re-scoping the progress dialog to preserve functionality of the existing stop button does not make sense.
   > Investigate suitable ways to minimally refactor existing progress bar behaviour such that GUI applications can, on a per-instance basis, control whether or not a "cancel" button is included in the dialog. AskUserQuestion if multiple candidate implementations are found. The functionality would therefore reside within cpp/gui/dialog/progress.* rather than cpp/gui/mrview/tool/screen_capture/.
4. > 1. Risk of Claude solution Dialog::ProgressBar::set_next_cancellable() is that developers may locally demote the command verbosity (App::log_level) to suppress dialogs in a specific context. A piece of code that sets that flag, then fails to spawn a dialog, whether due to not exceeding the timeout to be rendered or due to being locally deliberately disabled, may risk having that flag affect whatever non-suppressed progress dialog is next triggered in the application. This should not require diligence at the point of every invocation. Report on the probability of erroneous behaviour in this regard given Claude's chosen implementation.
   > 2. Report on why Claude proceeded with its chosen implementation despite explicit instruction from the user to execute AskUserQuestion in the presence of multiple solution candidates. One could debate whether AskUserQuestion should time out in some instances, but in response to explicit user request it should almost certainly not. Suggest modifications to user-level Claude configuration to mitigate the chance of this occurring again.
5.
6. > 1. Is it possible for Claude to write a hook that will detect the presence of text "AskUserQuestion" in a user prompt, and temporarily override variable CLAUDE_ASK_USER_TIMEOUT_MS, to provide a deterministic override of timeout rather than a model suggestion?
   > 2. Revert via git prior changes wherein the custom progress dialog for the screen capture tool was removed in favour of function set_next_cancellable().
   > 3. Investigate and comment on:
   >     3.1. Whether the desired optional availability of a cancel button in GUI progress dialogs can be achieved through class inheritance and type detection without necessitating any modification to the MR::ProgressBar class.
   >     3.2. A design where MR::ProgressBar has a latent member for cancel functionality availability that is inaccessible via any MR::ProgressBar constructor and always yields false but can then be overridden by a derivative class for GUI applications that opt in to the presence of a cancel button.
   >     3.3. The complexity involved in duplicating any current functionality of the mrview screen capture tool stop button whenever a dialog window cancel button is pressed.
7. > 3.1 is rejected on the basis that it can't be done with exactly zero modification to MR::ProgressBar. However 3.2 requires very explicit modification of MR::ProgressBar, with broadcast virtual capabilities not present in the base class. If MR::ProgressBar were to have minimal modification to make it polymorphic, then GUI apps requiring a cancel button could utilise a derivative class, and cpp/gui/dialog/progress.* could discover that through a dynamic_cast attempt. Would this not be superior encapsulation than 3.2? If so, attempt this implementation, ensuring that 3.3 is also satisfied.

Generated-by: Claude Sonnet 5 <noreply@anthropic.com>
