+++ cpp/gui/mrview/window.h
@@ -44,8 +44,6 @@ class CameraInteractor;
 
 namespace MR::GUI::MRView {
 
-class CaptureBuffer;
-
 class Window : public QMainWindow, ColourMapButtonObserver {
   Q_OBJECT
 
@@ -160,13 +158,11 @@ public:
 
   bool sync_focus_on() const { return sync_focus_action->isChecked(); }
 
-  //! Export the current visualisation to \a filepath.
-  /*! Off-screen rendering is used when \a supersample or \a msaa exceeds unity: the scene is rendered
-   *  off-screen at \a supersample times the display resolution, optionally with \a msaa-sample
-   *  anti-aliasing, leaving the visible window untouched; otherwise the on-screen framebuffer is
-   *  grabbed directly. The captured image is finally reduced by \a downsample, so the exported
-   *  resolution is the native resolution times \a supersample divided by \a downsample. */
-  void captureGL(const std::filesystem::path &filepath, int supersample = 1, int msaa = 1, int downsample = 1);
+  //! Render the current visualisation into the off-screen framebuffer already bound by the caller.
+  /*! Used by the screen-capture tool to render into its own off-screen buffer at \a supersample times
+   *  the display resolution: this only performs the scene render itself (scaling screen-space
+   *  annotations to match), leaving buffer allocation, anti-aliasing and read-back to the caller. */
+  void renderOffscreenGL(int supersample);
 
   //! The super-sampling ratio currently driving the render path (unity unless an off-screen capture is in progress).
   int supersample() const { return supersample_; }
@@ -262,9 +258,8 @@ private:
   GL::Lighting *lighting_;
   GL::Font font;
 
-  // Off-screen super-resolution screen capture (see captureGL()):
+  // Off-screen super-resolution screen capture (see renderOffscreenGL()):
   int supersample_ = 1;
-  std::unique_ptr<CaptureBuffer> capture_buffer;
   std::map<int, std::unique_ptr<GL::Font>> supersample_fonts;
   std::map<int, std::unique_ptr<ColourBars>> supersample_colourbars;
